Refactor out LocationData
